addAuthFirebase-and-firestore

+++ lib/widget/todo_widget.dart
@@ -3,12 +3,11 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_slidable/flutter_slidable.dart';
 import 'package:provider/provider.dart';
-import 'package:todo_app/provider/todos.dart';
 import 'package:todo_app/utils.dart';
 
 import '../model/todo.dart';
 import '../page/edit_todo_page.dart';
-import '../provider/shared_preferences_todos_provider.dart';
+import '../provider/todos_provider.dart';
 
 
 class TodoWidget extends StatelessWidget {
@@ -49,7 +48,33 @@ class TodoWidget extends StatelessWidget {
                 backgroundColor: Colors.red,
                 icon: Icons.delete_rounded,
                 onPressed: (context) { 
-                  deleteTodo(context,todo);
+                  
+                     showDialog(context: context,
+                 builder: (context) =>AlertDialog(
+                  title:Text("Warning!!!"),
+                    content: Text("Are you sure to delete?" , style: TextStyle(color: Colors.red),), 
+                    actions: [
+                      TextButton(child: Text('CANCEL'),
+                       onPressed: () => Navigator.of(context).pop() ,
+                      ),
+                       TextButton(child: Text('Delete' , style: TextStyle(color: Colors.red)),
+                       onPressed: () {
+                         deleteTodo(context,todo);
+                        Navigator.of(context).pop() ;
+                        
+                 
+               } 
+               
+              ),
+                
+                    ],  
+                 )
+                 );
+                // fina
+
+
+
+                 
                 },
                 
               ),
@@ -75,7 +100,7 @@ class TodoWidget extends StatelessWidget {
                 checkColor: Colors.white,
                 value: todo.isDone, 
                 onChanged: (_){
-                  final provider = Provider.of<SharedPreferencesTodosProvider>(context,listen: false);
+                  final provider = Provider.of<TodosProvider>(context,listen: false);
                   final isDone= provider.toggleTodoStatus(todo);
                   Utils.showSnackBar(context, 
                    isDone ? 'Task completed' :'Task marked uncompleted',
@@ -112,7 +137,7 @@ class TodoWidget extends StatelessWidget {
 }
 
   deleteTodo(BuildContext context, Todo todo) {
-    final provider = Provider.of<SharedPreferencesTodosProvider>(context, listen: false);
+    final provider = Provider.of<TodosProvider>(context, listen: false);
     provider.removeTodo(todo);
 
     Utils.showSnackBar(context,'Deleted the Task');
